Makefile: initial commit
This Makefile simplifies building, debugging with clang and clang-tidy,
cleaning, and running the output, as well as producing the devlog. It
was designed with automatic dependency detection and ease of future
expansion in mind.

Hiding the echo commands through boolean short-circuiting makes debug
info easier to read, but make the output a bit more confusing.